fixes, refactoring
