+++ src/components/shopping-cart-item/index.js
@@ -1,0 +1,3 @@
+import ShoppingCartItem from "./shopping-cart-item";
+
+export default ShoppingCartItem;
