+++ src/services/shop-service.js
@@ -47,11 +47,9 @@ export default class ShopService {
         const response = await fetch(`${this._apiUrl}/cart/${productId}/reduce`, {
             method: 'PUT'
         })
-        .then(response => {
-            if (!response.ok) {
-                throw new Error(`Could not fetch ${this._apiUrl}/cart/${productId}/reduce, received ${response.status}`);
-            }
-        });
+        if (!response.ok) {
+            throw new Error(`Could not fetch ${this._apiUrl}/cart/${productId}/reduce, received ${response.status}`);
+        }
         
         return response;
     }
@@ -60,11 +58,9 @@ export default class ShopService {
         const response = await fetch(`${this._apiUrl}/cart/${productId}`, {
             method: 'DELETE'
         })
-        .then(response => {
-            if (!response.ok) {
-                throw new Error(`Could not fetch ${this._apiUrl}/cart/${productId}, received ${response.status}`);
-            }
-        });
+        if (!response.ok) {
+            throw new Error(`Could not fetch ${this._apiUrl}/cart/${productId}, received ${response.status}`);
+        }
         
         return response;
     }
